+++ changelog.md
@@ -1,1 +1,1 @@
-- EnhancedBlockEntities added. As there is currently no update for 1.20.4 available. As soon as an official update is available, this will be removed again.
+- There is now a function that summarizes duplicate messages.